==> program/1_preprocess.ipynb <==
文本预处理，将所有训练数据合并成一个文件，去掉标点，停词

import re,os
import numpy as np
import pandas as pd
np.random.seed(42)

INPUT_PATH = '../input/'
CACHE_PATH = '../cache/'
OUTPUT_PATH ='../output'

# 读取数据
df_train_first = pd.read_csv(INPUT_PATH + 'train_first.csv')
df_train_second = pd.read_csv(INPUT_PATH + 'train_second.csv')
df_predict_first = pd.read_csv(INPUT_PATH + 'predict_first.csv')
df_predict_second = pd.read_csv(INPUT_PATH + 'predict_second.csv')

# 合并数据
df_train = pd.concat([df_train_first,df_train_second,df_predict_first],axis=0,ignore_index=True)
df_predict = df_predict_second.copy()

import os
LTP_DATA_DIR = '../ltp_data_v3.4.0/'  # ltp模型目录的路径
cws_model_path = os.path.join(LTP_DATA_DIR, 'cws.model')  # 分词模型路径，模型名称为`cws.model`

from pyltp import Segmentor
segmentor = Segmentor()  # 初始化实例
segmentor.load_with_lexicon(cws_model_path, './china_position')  # 加载模型，第二个参数是您的外部词典文件路径


def hgda_nlp(text):

    words = segmentor.segment(text)
    # segmentor.release()
    return words


# 预处理单个词
def preprocess_word(word):
    # 去掉标点
    word = word.strip('\'"?!,.():;？！，。…“”（）：；<>《》/ 、rn【】[]|~#%&*br')
    # 去掉 - &
    word = re.sub(r'(-|\')', '', word)
    
    return word

# 预处理句子
def preprocess_sentence(sentence, stop_words=False):
    sentence = sentence.lower()
    # 转化URL
    sentence = re.sub(r'((www\.[\S]+)|(https?://[\S]+))', ' URL ', sentence)
    # 去掉空白，单双引号
    sentence = sentence.strip(' "\'')
    # 分词
    words = hgda_nlp(sentence)
    # 分别处理每一个单词
    sentence = []
    for word in words:
        word = preprocess_word(word)
        if len(word) > 0:
            sentence.append(word)
    if len(sentence) == 0:
        return ' '.join(list('很好'))
    else:
        return ' '.join(sentence)

# 拷贝数据
df_train_processed = df_train.copy()
df_predict_processed = df_predict.copy()

# 预处理文本
df_train_processed['Discuss'] = df_train['Discuss'].map(preprocess_sentence)
df_predict_processed['Discuss'] = df_predict['Discuss'].map(preprocess_sentence)

df_train_processed

# 保存预处理后的数据集
df_train_processed.to_csv(CACHE_PATH + 'train_processed_all.csv',index=False)
df_predict_processed.to_csv(CACHE_PATH+ 'predict_processed_all.csv',index=False)
