==> program/2_word2vec.ipynb <==
import os
import numpy as np
import pandas as pd
from gensim.models.word2vec import Word2Vec
from gensim.models import word2vec

INPUT_PATH = '../input/'
CACHE_PATH = '../cache/'
OUTPUT_PATH ='../output/'

# 读取数据
df_train = pd.read_csv(CACHE_PATH + 'train_processed_all.csv')
df_predict = pd.read_csv(CACHE_PATH + 'predict_processed_all.csv')

EMBEDDING_SIZE = 300
EMBEDDING_FILE_NAME = CACHE_PATH  + 'chinese_vector_300.txt'
CORPUS_FILE_NAME = CACHE_PATH  + 'corpus_all.txt'

# 获取语库
corpus = np.concatenate([df_train['Discuss'],df_predict['Discuss']])

# 保存文本集
with open(CORPUS_FILE_NAME,'w') as file_vec:
    for line in corpus:
        file_vec.writelines(line)

# 训练word2vec
sentences = word2vec.Text8Corpus(CORPUS_FILE_NAME)
model = Word2Vec(sentences, min_count=20, size= EMBEDDING_SIZE, window=6)
model.wv.save_word2vec_format(EMBEDDING_FILE_NAME, binary=False)

model.most_similar_cosmul('北京')

len(model['北'])
